Added new functionality to Repository and Student.
Added new dictionaries to Repository to store lists of required and elective courses for specific majors. Added lists to Student to store required courses remaining and electives offered for their major if they have not completed and passed at least one elective. Also created new methods to parse the data from a text file and store in the new dictionaries and lists, and to check that the student has completed a course with a passing grade, update the student prettytable, and create a majors pretty table. HW09KyleRozanitisUnitTests.py now contains unit tests.